Flags

NFFT3.PRE_PHI_HUT — Constant

PRE_PHI_HUT

precompute and store values $\hat{\phi}(k)$ of the Fourier transform of the window function $\hat{\phi}$.

NFFT3.FG_PSI — Constant

FG_PSI

fast Gaussian gridding.

NFFT3.PRE_LIN_PSI — Constant

PRE_LIN_PSI

linear interpolation of the window function from a lookup table.

NFFT3.PRE_FG_PSI — Constant

PRE_FG_PSI

fast Gaussian gridding.

NFFT3.PRE_PSI — Constant

PRE_PSI

precomputation based on tensor product structure of the window function.

NFFT3.PRE_FULL_PSI — Constant

PRE_FULL_PSI

calculate and store all values $\tilde{\psi}(x_j - \frac{1}{n} \odot l)$.

NFFT3.MALLOC_X — Constant

MALLOC_X

allocate memory for node $x_j$.

NFFT3.MALLOC_F_HAT — Constant

MALLOC_F_HAT

allocate memory for coefficient$\hat{f}_k$.

NFFT3.MALLOC_F — Constant

MALLOC_F

allocate memory for approximate function value $f_j$.

NFFT3.FFT_OUT_OF_PLACE — Constant

FFT_OUT_OF_PLACE

FFTW uses disjoint input/output vector.

NFFT3.FFTW_INIT — Constant

FFTW_INIT

initialize FFTW plan.

NFFT3.NFFT_SORT_NODES — Constant

NFFT_SORT_NODES

internal sorting of the nodes $x_j$ that may increase performance.

NFFT3.NFFT_OMP_BLOCKWISE_ADJOINT — Constant

NFFT_OMP_BLOCKWISE_ADJOINT

blockwise calculation for adjoint NFFT in the case of OpenMP.

NFFT3.NFCT_SORT_NODES — Constant

NFCT_SORT_NODES

internal sorting of the nodes $x_j$ that may increase performance.

NFFT3.NFCT_OMP_BLOCKWISE_ADJOINT — Constant

NFCT_OMP_BLOCKWISE_ADJOINT

blockwise calculation for adjoint NFFT in the case of OpenMP.

NFFT3.NFST_SORT_NODES — Constant

NFST_SORT_NODES

internal sorting of the nodes $x_j$ that may increase performance.

NFFT3.NFST_OMP_BLOCKWISE_ADJOINT — Constant

NFST_OMP_BLOCKWISE_ADJOINT

blockwise calculation for adjoint NFFT in the case of OpenMP.

NFFT3.FFTW_MEASURE — Constant

FFTW_MEASURE

find optimal plan by executing several FFTs and compare times.

NFFT3.FFTW_DESTROY_INPUT — Constant

FFTW_DESTROY_INPUT

an out-of-place transform is allowed to overwrite the input array with arbitrary data.

NFFT3.FFTW_UNALIGNED — Constant

FFTW_UNALIGNED

the algorithm may not impose any unusual alignment requirements on the input/output arrays (not necessary in most context).

NFFT3.FFTW_CONSERVE_MEMORY — Constant

FFTW_CONSERVE_MEMORY

conserving memory.

NFFT3.FFTW_EXHAUSTIVE — Constant

FFTW_EXHAUSTIVE

behaves like FFTW_PATIENT with an even wider range of tests.

NFFT3.FFTW_PRESERVE_INPUT — Constant

FFTW_PRESERVE_INPUT

input vector is preserved and unchanged.

NFFT3.FFTW_PATIENT — Constant

FFTW_PATIENT

behaves like FFTW_MEASURE with a wider range of tests.

NFFT3.FFTW_ESTIMATE — Constant

FFTW_ESTIMATE

use simple heuristic instead of measurements to pick a plan.

NFFT3.FFTW_WISDOM_ONLY — Constant

FFTW_WISDOM_ONLY

a plan is only created if wisdom from tests is available.