Nonequispaced Fast Cosine Transform

NFFT3.NFCT — Type

NFCT{D}

A NFCT (Nonequispaced fast cosine transform) plan, where D is the dimension.

The NFCT realizes a direct and fast computation of the discrete nonequispaced cosine transform. The aim is to compute

\[f^c (x) \colon = \sum_{k \in I_N} \hat{f}^{c}_{k} \cos (k \dot x), \quad x \in \mathbb{R}^D\]

at given (nonequidistant) knots $x_k \in [0, \pi ]^D, \; k = 0, \cdots, M-1$, coefficients $\hat{f}^{c}_{k} \in \mathbb{R}$, $k \in I_N \colon = \{ k \in \mathbb{Z}^{D} \colon 0 \leq k_i \leq N_i, \, \forall i = 1, \cdots, D\}$ for some multibandlimit vector $N \in \mathbb{N}^{D}$. The transposed (adjoined) problem reads as

\[\hat{h}_k \colon = \sum_{j \in I_M} f_j \cos (k \dot x_j), \quad k \in I^{D}_N\]

for given knots $x_k \in [0, \pi ]^D, \; k = 0, \cdots, M-1$, and coefficients $f_j \in \mathbb{C}, j \in I_M^l$.

Fields

N - the multibandlimit of the trigonometric polynomial f.

M - the number of nodes.

n - the oversampling per dimension.

m - the window size. Larger m means more accuracy but also more computational costs.

f1 - the NFCT flags.

f2 - the FFTW flags.

init_done - indicates if the plan is initialized.

finalized - indicates if the plan is finalized.

x - the nodes.

f - the function values.

fhat - the Fourier coefficients.

plan

Constructor

NFCT{D}(N::NTuple{D,Int32},M::Int32,n::NTuple{D,Int32},m::Int32,f1::UInt32,f2::UInt32) where {D}

Additional Constructor

NFCT(N::NTuple{D,Int32},M::Int32,n::NTuple{D,Int32},m::Int32,f1::UInt32,f2::UInt32) where {D}
NFCT(N::NTuple{D,Int32},M::Int32) where {D}

See also

NFFT

NFCT

The NFCT (Nonequispaced fast cosine transform) ([Plonka, Potts, Steidl, Tasche, 2018], Sec. 7.4) realizes a multivariate fast cosine transform for nonequispaced knots. The aim is to compute

\[ f^c (x) \coloneqq \sum_{ k \in I_{ N}} \hat{f}_{ k}^c \cos( k \cdot x), \quad x \in \mathbb{R}^d\]

at given (nonequidistant) knots $x_{k} \in [ 0,\pi ]^d, \, k = 0, \ldots, M-1$, coefficients $\hat{f_{ k}^c} \in \mathbb{R}, \, k \in I_{ N} \coloneqq \{ k\in \mathbb{N}^d:\, 0 \leq k_i \leq N_i \; \forall \, i = 1, \ldots, d \}$ for some multibandlimit vector $N \in \mathbb{N}^d$.
For simplicity, we only consider the univariate case ($d=1$). The multivariate case, which is also implemented in the module, can be derived analogously from the NFFT algorithm (see e.g. [Schmischke, 2018] or [Plonka, Potts, Steidl, Tasche, 2018], Chapter 7). Hence, we want to approximate

\[ f^c \colon [ 0, \pi ] \to \mathbb{C}, \quad f^c(x_j) = \sum_{k=0}^{N-1} \hat{f}^{c}_k \cos(k x_j) = \sum_{k=-N}^{N-1} \hat{f}_k \mathrm{e}^{ \mathrm{i} k x_j},\]

with $x_j \in [0,\pi], \, j=0, \ldots, M-1$ the nodes and $\hat{f}_k^c \in \mathbb{R}, \, k = 0, \ldots, N-1$ the coefficients. Observe that the second equality sign is only justified when $\hat{f}_0 = \hat{f}_0^c, \, \hat{f}_{_N}=0, \, \hat{f}_k = \hat{f}_{-k}= \frac{1}{2}\hat{f}_k^c$. Similarly to the NFFT, we choose $g_{\ell} \in \mathbb{R}$ in the following term such that $s_1$ with a given window function $\varphi\in L_2(\mathbb{R})\cap L_1(\mathbb{R})$ approximates $f$

\[ s_1(x) \coloneqq \sum_{\ell=0}^{\sigma N} g_{\ell} \ \tilde{\varphi}(x-\frac{\pi\ell}{\sigma N}), \quad x\in \mathbb{R}.\]

$\sigma \in \mathbb{N}$ is an oversampling factor fulfilling $\sigma N \in \mathbb{N}$. Replacing $\varphi$ by its Fourier series $\varphi (x) = \sum_{k\in \mathbb{N}} \hat{g}_k \mathrm{e}^{\mathrm{i} k x}$, comparing the definition of $s_1(\cdot)$ and $f^c(\cdot)$ and taking the symmetry into account, we obtain

\[ g_{\ell} = \frac{1}{\sigma N} \sum_{k=0}^{\sigma N} (\varepsilon_{\sigma N}(k))^2 \hat{g}_k \cos ( \frac{2 \pi k \ell}{\sigma N} ), \quad \ell = 0, \ldots, \sigma N.\]

Here, $\varepsilon_{\sigma N}(j)\coloneqq 1, \, j=1,\ldots,\sigma N-1$, and $\varepsilon_{\sigma N}(0)\coloneqq \varepsilon_{\sigma N}(\sigma N)\coloneqq \frac{\sqrt{2}}{2}$. Now, we can compute $g_{\ell}$ by an DCT-I of length $\sigma N +1$ and obtain $f(x) \approx s_1(x)$.
Next, we want to replace $\varphi$ by its truncation

\[ \psi(x) = \varphi(x) \chi_Q (x) = \begin{cases} \varphi(x) & x\in Q\coloneqq[ -\frac{2\pi m}{\sigma N},\frac{2\pi m}{\sigma N} ] , \\ 0, & else. \end{cases}\]

As in the definition of $s_1(\cdot)$, we replace $\psi$ by its periodization $\tilde{\psi}$ and define

\[ s(x) \coloneqq \sum_{\ell=\lfloor 2\sigma N x\rfloor -m}^{\lceil 2\sigma N x\rceil +m} g_\ell \ \tilde{\psi}(x-\frac{\pi \ell}{\sigma N}),\quad x\in\mathbb{R}\]

We arrive at $f(x)\approx s_1(x) \approx s(x)$.

Pseudocode

Input: $N, M \in \mathbb{N}, \, \sigma > 1, \, m \in \mathbb{N}, \, x_j \in [0, \pi] \text{ for } j = 0, \ldots, M - 1, \, f^{c}_k \in \mathbb{R} \text{ for } k = 0, \ldots, N - 1.$

Precomputation: Compute the nonzero Fourier coefficients $c_k(\tilde{\varphi}) \text{ for all } k = 0, \ldots, N - 1.$ Compute the values $\tilde{\psi} (x_j -\frac{\pi \ell}{\sigma N}) \text{ for } j = 0, \ldots, M - 1 \text{ and } \ell \in I_{\sigma N, \ m} (x_j).$

Set $\hat{g}_k \coloneqq \begin{cases} \frac{\hat{f}^c_k}{2 (\varepsilon_{\sigma N}(k))^2 c_k(\tilde{\varphi})} & k = 0, \ldots, N - 1, \\ 0 & k = N, \ldots, \sigma N .\end{cases}$

Compute $g_{\ell} = \frac{1}{\sigma N} \sum_{k=0}^{\sigma N} (\varepsilon_{\sigma N}(k))^2 \hat{g}_k \cos ( \frac{\pi k \ell}{\sigma N} ), \quad \ell = 0, \ldots, \sigma N$ using a fast algorithm of DCT-I($\sigma N + 1$), see Algorithm 6.28 or 6.35 in [Plonka, Potts, Steidl, Tasche, 2018].

Compute $s(x_j) \coloneqq \sum_{\ell=\lfloor 2\sigma N x_j \rfloor -m}^{\lceil 2 \sigma N x_j \rceil +m} g_\ell\ \tilde{\psi}(x_j -\frac{\pi \ell}{\sigma N}), \quad j = 0, \ldots, M-1$.

Output: $s(x_j), \, j = 0, \ldots, M-1$, approximate values of $f^c(x_j)$.

Computational cost: $\mathcal{O}(N \log{N} + m \ M)$

Transposed problem

The transposed problem reads as

\[ h(k) \coloneqq \sum_{ j \in I_M^l} f_{j}\ \cos( k \ x_j}), \quad k \in I_{ N}^d \coloneqq \{ k \in \mathbb{N}^d \colon 0 \leq k_i \leq N_i \}\]

for given knots ${x}_k \in [ 0,\pi ]^d, \, k=0,\ldots,M-1$, and coefficients $f_j \in \mathbb{C}, j \in I_M^l$. The algorithm for the fast evaluation of

\[ h(k) \coloneqq \sum^{M-1}_{ j = 0} h_{j} \ \cos(k \ x_j), \quad k = 0, \ldots, N-1,\]

with nonequispaced nodes ${x}_j \in [ 0,\pi ], \, j=0,\ldots,M-1,$ can easily be derived from the duality of the two problems.

Functions

NFFT3.nfct_finalize_plan — Function

nfct_finalize_plan(P::NFCT{D})

destroys a NFCT plan structure.

Input

P - a NFCT plan structure.

See also

NFCT{D}, nfct_init

NFFT3.nfct_init — Function

nfct_init(P)

intialises a transform plan.

Input

P - a NFCT plan structure.

See also

NFCT{D}, nfct_finalize_plan

NFFT3.nfct_trafo — Function

nfct_trafo(P)

computes a NFCT.

Input

P - a NFCT plan structure.

See also

NFCT{D}, trafo_direct

NFFT3.nfct_adjoint — Function

nfct_adjoint(P)

computes an adjoint NFCT.

Input

P - a NFCT plan structure.

See also

NFCT{D}, adjoint_direct

NFFT3.nfct_trafo_direct — Function

nfct_trafo_direct(P)

computes a NFCT.

Input

P - a NFCT plan structure.

See also

NFCT{D}, nfct_trafo

NFFT3.nfct_adjoint_direct — Function

nfct_adjoint_direct(P)

computes an adjoint NFCT.

Input

P - a NFCT plan structure.

See also

NFCT{D}, adjoint

Literature

[Plonka, Potts, Steidl, Tasche, 2018] G. Plonka, D. Potts, G. Steidl and M. Tasche. Numerical Fourier Analysis: Theory and Applications. Springer Nature Switzerland AG, 2018. doi: 10.1007/978-3-030-04306-3.

[Schmischke, 2018] M. Schmischke. Nonequispaced Fast Fourier Transform (NFFT) Interface for Julia. 2018. arXiv: 1512.02814.