Nonequispaced Fast Sine Transform

NFFT3.NFST — Type

NFST{D}

A NFST (Nonequispaced fast sine transform) plan, where D is the dimension.

The NFST realizes a direct and fast computation of the discrete nonequispaced sine transform. The aim is to compute

\[f^c (x) \colon = \sum_{k \in I_N} \hat{f}^{c}_{k} \sin (k \dot x), \quad x \in \mathbb{R}^D\]

at given (nonequidistant) knots $x_k \in [0, \pi ]^D, \; k = 0, \cdots, M-1$, coefficients $\hat{f}^{c}_{k} \in \mathbb{R}$, $k \in I_N \colon = \{ k \in \mathbb{Z}^{D} \colon 0 \leq k_i \leq N_i, \, \forall i = 1, \cdots, D\}$ for some multibandlimit vector $N \in \mathbb{N}^{D}$. The transposed (adjoined) problem reads as

\[\hat{h}_k \colon = \sum^{M-1}_{j = 0} f_j e^{-2 \pi \mathrm{i} \mathbf{k} \cdot \mathbf{x_j}}, \; k \in I_N\]

for given knots $x_k \in [0, \pi ]^D, \; k = 0, \cdots, M-1$, and coefficients fj ∈ C, j ∈ I l M.

Fields

N - the multibandlimit of the trigonometric polynomial f.

M - the number of nodes.

n - the oversampling per dimension.

m - the window size. Larger m means more accuracy but also more computational costs.

f1 - the NFST flags.

f2 - the FFTW flags.

init_done - indicates if the plan is initialized.

finalized - indicates if the plan is finalized.

x - the nodes.

f - the function values.

fhat - the Fourier coefficients.

plan

Constructor

NFST{D}(N::NTuple{D,Int32},M::Int32,n::NTuple{D,Int32},m::Int32,f1::UInt32,f2::UInt32) where {D}

Additional Constructor

NFST(N::NTuple{D,Int32},M::Int32,n::NTuple{D,Int32},m::Int32,f1::UInt32,f2::UInt32) where {D}
NFST(N::NTuple{D,Int32},M::Int32) where {D}

See also

NFFT

NFST

We modify the NFFT in order to derive a fast algorithm for the evaluation of the odd, $2 \pi$-periodic trigonometric polynomial

\[ f^s (x) \coloneqq \sum^{N-1}_{k=1} \hat{f}_{k}^s \sin( k \ x), \quad x \in \mathbb{R}\]

at nonequispaced nodes $x_j \in (0,\pi)$. To this end, we rewrite $f^s$ as a sum of exponentials and obtain

\[ \mathrm{i} \ f^s (x) = f(x) = \sum^{N-1}_{k=-N} \hat{f}_{k} \mathrm{e}^{\mathrm{i} \ k \ x} = \mathrm{i} \sum^{N-1}_{k=1} 2 \hat{f}_{k} \sin( k \ x), \quad x \in \mathbb{R}\]

with $\hat{f}_0 = \hat{f}_{-N} = 0$ and $\hat{f}_k = -\hat{f}_{-k} = \frac{1}{2}\hat{f}^{s}_k$ for $k = 1, \ldots, N - 1$. Similarly as before, we approximate $f(x)$ by a function $s_1(x)$ and obtain for the coefficients $g_{\ell}$ for $\ell = 1, \ldots, \sigma N - 1$

\[ - \mathrm{i} g_{\ell} = \frac{- \mathrm{i}}{2 \sigma N} \sum^{\sigma N - 1}_{k=-\sigma N} \hat{g}_{k} \mathrm{e}^{\pi \ \mathrm{i} \ k \ \ell / (\sigma N)} = \frac{1}{\sigma N} \sum^{\sigma N - 1}_{k=1} \hat{g}_{k} \sin( \frac{\pi \ k \ \ell}{\sigma N})\]

and particularly $g_0 = g_{\sigma N} = 0$. Moreover, we observe that $g_{2 \sigma N r - \ell} = -g_{\ell}$ for all $r \in \mathbb{Z}$. Finally we compute the sum

\[ \mathrm{i} \ s(x_j) \coloneqq \sum_{\ell = \lfloor 2 \sigma N x \rfloor - m }^{\lceil 2 \sigma N x \rceil + m} \mathrm{i} \ g_{\ell} \ \tilde{\psi}(x_j - \frac{\pi \ell}{\sigma N})\]

and obtain the approximate values of $f^s(x_j) = \mathrm{i} \ f(x_j) \approx \mathrm{i} \ s(x_j), \; j = 0, \ldot,M - 1$.

Pseudocode

Input: $N, M \in \mathbb{N}, \, \sigma > 1, \, m \in \mathbb{N}, \, x_j \in (0, \pi) \text{ for } j = 0, \ldots, M - 1, \, f^{s}_k \in \mathbb{R} \text{ for } k = 0, \ldots, N - 1.$

Precomputation: Compute the nonzero Fourier coefficients $c_k(\tilde{\varphi}) \text{ for all } k = 0, \ldots, N - 1.$ Compute the values $\tilde{\psi} (x_j -\frac{\pi \ell}{\sigma N}) \text{ for } j = 0, \ldots, M - 1 \text{ and } \ell \in I^{\mathrm{T}}_{\sigma N, \ m} (x_j).$

Set $\hat{g}_k \coloneqq \begin{cases} \frac{\hat{f}^{s}_k}{2 c_k(\tilde{\varphi})} & k = 1, \ldots, N - 1, \\ 0 & k = 0 \text{ and } k = N, \ldots, \sigma N .\end{cases}$

Compute $g_{\ell} = \frac{1}{\sigma N} \sum_{k=1}^{\sigma N - 1} \hat{g}_k \sin ( \frac{\pi k \ell}{\sigma N} ), \quad \ell = 1, \ldots, \sigma N - 1$ using a fast algorithm of DST-I($\sigma N - 1$), see Table 6.1 and Remark 6.40 in [Plonka, Potts, Steidl, Tasche, 2018] and set $g_0 \coloneqq 0$.

Compute $s(x_j) \coloneqq \sum_{\ell=\lfloor 2\sigma N x_j \rfloor -m}^{\lceil 2 \sigma N x_j \rceil +m} g_\ell\ \tilde{\psi}(x_j -\frac{\pi \ell}{\sigma N}), \quad j = 0, \ldots, M-1$.

Output: $s(x_j), \, j = 0, \ldots, M-1$, approximate values of $f^s(x_j)$.

Computational cost: $\mathcal{O}(N \log{N} + m \ M)$

Transposed problem

The transposed problem reads as

\[ h(k) \coloneqq \sum_{ j \in I_M^l} f_{j}\ \sin( k \ x_j}), \quad k \in I_{ N}^d \coloneqq \{ k \in \mathbb{N}^d \colon 0 \leq k_i \leq N_i \}\]

for given knots ${x}_k \in [ 0,\pi ]^d, \, k=0,\ldots,M-1$, and coefficients $f_j \in \mathbb{C}, j \in I_M^l$. The algorithm for the fast evaluation of

\[ h(k) \coloneqq \sum^{M-1}_{ j = 0} h_{j} \ \sin(k \ x_j), \quad k = 0, \ldots, N-1,\]

with nonequispaced nodes ${x}_j \in [ 0,\pi ], \, j=0,\ldots,M-1,$ can easily be derived from the duality of the two problems.

Functions

NFFT3.nfst_finalize_plan — Function

nfst_finalize_plan(P)

destroys a NFST plan structure.

Input

P - a NFST plan structure.

See also

NFST{D}, nfst_init

NFFT3.nfst_init — Function

nfst_init(P)

intialises a transform plan.

Input

p - a NFST plan structure.

See also

NFST{D}, nfst_finalize_plan

NFFT3.nfst_trafo — Function

nfst_trafo(P)

computes a NFST.

Input

P - a NFST plan structure.

See also

NFST{D}, nfst_trafo_direct

NFFT3.nfst_adjoint — Function

nfst_adjoint(P)

computes an adjoint NFST.

Input

P - a NFST plan structure.

See also

NFST{D}, nfst_adjoint_direct

NFFT3.nfst_trafo_direct — Function

nfst_trafo_direct(P)

computes a NFST.

Input

P - a NFST plan structure.

See also

NFST{D}, nfst_trafo

NFFT3.nfst_adjoint_direct — Function

nfst_adjoint_direct(P)

computes an adjoint NFST.

Input

P - a NFST plan structure.

See also

NFST{D}, nfst_adjoint