Nonequispaced Fast Fourier Transform

NFFT3.NFFT — Type

NFFT{D}

A NFFT (nonequispaced fast Fourier transform) plan, where D is the dimension.

The classic FFT (Fast Fourier transform) algorithm computes the discrete Fourier transform

\[f_j \colon = \sum^{\frac{N}{2} - 1}_{k = - \frac{N}{2}} \hat{f}_k e^{-2 \pi \mathrm{i} \frac{k_j}{N}}\]

for $j = - \frac{N}{2}, \dots, \frac{N}{2} - 1$ and given complex coefficients $\hat{f}_{k} \in \mathbb{C}$. Using a divide and conquer approach, the number of floating point operations is reduced from $\mathcal {O}(N^2)$ for a straightforward computation to only $\mathcal {O}(N \log N)$. However, two shortcomings of traditional schemes are the need for equispaced sampling and the restriction to the system of complex exponential functions. The NFFT overcomes the need for an equispaced sampling grid. Considering a D-dimensional trigonometric polynomial

\[f \colon \mathbb{T}^D \to \mathbb{C}, \; f(x) \colon = \sum_{k \in I_{N}} \hat{f}_k e^{-2 \pi \mathrm{i} \mathbf{k} \cdot \mathbf{x}}\]

with an index set $I_N \colon = \{ k \in \mathbb{Z}^{D} \colon - \frac{N_i}{2} \leq k_i \leq \frac{N_i}{2} - 1, \, i = 0, \cdots, D-1 \}$ where $N \in 2 \mathbb{N}^{D}$ is the multibandlimit. The NDFT (non uniform discrete fourier transform) is its evaluation at $M \in 2 \mathbb{N}$ nonequispaced points $x_j \in \mathbb{T}^D$ for $j = 0, 1, \cdots, M$,

\[f(x_j) \colon = \sum_{k \in I_{N}} \hat{f}_k e^{-2 \pi \mathrm{i} \mathbf{k} \cdot \mathbf{x_j}}\]

with given coefficients $\hat{f}_k \in \mathbb{C}$ where we identify the smooth manifold of the torus $\mathbb{T}$ with $[−1/2, 1/2)$. The NFFT is an algorithm for the fast evaluation of the NDFT and the adjoint problem, the fast evaluation of the adjoint NDFT

\[\hat{h}_k \colon = \sum^{M-1}_{j = 0} f_j e^{-2 \pi \mathrm{i} \mathbf{k} \cdot \mathbf{x_j}}, \; k \in I_N\]

for given coefficients $f_j \in \mathbb{C}$. In general, the adjoint NDFT is not the inverse transform of the NDFT. The available NFFT3 library [KeinerKunisPotts] provides C routines for the NFFT (for computing the NDFT in one or more dimensions, of arbitrary input size, and of complex data), applications such as the fast evaluation of sums

\[g(y_j) \colon = \sum^{N}_{k = 1} \alpha_k K(\lVert y_j - x_k \rVert_2), \; j = 1, \dots, M\]

for given coefficients $\alpha_k \in \mathbb{C}$, nodes $x_k, y_j \in \mathbb{R}^D$ and a radial kernel function $K \colon [0, \infty) \to [0, \infty)$, and generalizations such as the NNFFT for nonequispaced nodes in time and frequency domain.

Fields

N - the multibandlimit.

M - the number of nodes.

n - the oversampling per dimension.

m - the window size. Larger m means more accuracy but also more computational costs.

f1 - the NFFT flags.

f2 - the FFTW flags.

init_done - indicates if the plan is initialized.

finalized - indicates if the plan is finalized.

x - the nodes.

f - the function values.

fhat - the Fourier coefficients.

plan - plan (C pointer).

Constructor

NFFT{D}(N::NTuple{D,Int32},M::Int32,n::NTuple{D,Int32},m::Int32,f1::UInt32,f2::UInt32) where D

Additional Constructor

NFFT(N::NTuple{D,Int32},M::Int32,n::NTuple{D,Int32},m::Int32,f1::UInt32,f2::UInt32) where {D}
NFFT(N::NTuple{D,Int32},M::Int32) where {D}

NFFT

The nonequispaced fast Fourier transform Keiner, Kunis, Potts, 2006 overcomes one of the main shortcomings of the FFT - the need for an equispaced sampling grid. Considering the evaluation of the $d$-dimensional trigonometric polynomial

\[ f \colon \mathbb{T}^d \to \mathbb{C}, \; \pmb{x} \mapsto \sum_{\pmb{k} \in I_{N}} \hat{f}_{\pmb{k}} \mathrm{e}^{-2 \pi \mathrm{i} \ \pmb{k} \cdot \pmb{x}}\]

with multibandlimit $\pmb{N} \in 2 \mathbb{N}^d$ and index set

\[ I_{N} \coloneqq \{ \pmb{k} \in \mathbb{Z}^d: - \frac{N_i}{2} \leq k_i \leq \frac{N_i}{2} - 1 \; \forall i \in \{ 1,2,\ldots, d \} \}.\]

The first approximation is a linear combination of a shifted periodized window function $\tilde{\varphi}$

\[ s_1(\pmb{x}) = \sum_{\pmb{\ell} \in I_{\pmb{n}}} g_{\pmb{\ell}} \tilde{\varphi} \left( \pmb{x} - \frac{1}{\pmb{n}} \odot \pmb{\ell} \right)\]

where $\frac{1}{\pmb{n}}$ is the elementwise inversion of the vector $\pmb{n}$. We choose an oversampling vector $\pmb{\sigma} > 1$ componentwise and obtain the index set by

\[ \pmb{n} \coloneqq \pmb{\sigma} \odot \pmb{N}.\]

Here, $\odot$ denotes the componentwise product. Note that one could skip the choice of $\pmb{\sigma}$ entirely and directly choose $n_i > N_i$ for $i= 1, 2, \ldots,d$. The standard choice in the C library is at the moment

\[ n_i = 2^{\lceil \log_2 N_i \rceil + 1}, i= 1, 2, \ldots, d. \]

If the window function $\varphi: \R^d \to \R$ has a one-periodization

\[ \tilde{\varphi}(\pmb{x}) = \sum_{\pmb{r} \in \mathbb{Z}^d} \varphi(\pmb{x}+\pmb{r}) \]

with a uniformly convergent Fourier series, we use the Poisson summation formula and obtain the Fourier coefficients

\[ c_{\pmb{k}}(\tilde{\varphi}) = \hat{\varphi}(\pmb{k}).\]

Here, $\hat{\varphi}$ is the Fourier transform of $\varphi$ defined by

\[ \hat{\varphi}(\pmb{k}) = \int_{\mathbb{T}^d} \varphi(\pmb{x}) \mathrm{e}^{-2 \pi \mathrm{i} \ \pmb{k} \cdot \pmb{x}} \mathrm{d} \pmb{x}.\]

Replacing $\tilde{\varphi}$ by its Fourier series and splitting the sum in $s_1(x)$ yields

\[\begin{aligned} s_1(\pmb{x}) & = \sum_{\pmb{\ell} \in I_{\pmb{n}}} g_{\pmb{\ell}} \sum_{\pmb{k} \in \mathbb{Z}^d} c_{\pmb{k}}(\tilde{\varphi}) \mathrm{e}^{-2 \pi \mathrm{i} \ \pmb{k} \cdot \left( \pmb{x} - \frac{1}{\pmb{n}} \odot \pmb{\ell}\right) } \\ & = \sum_{\pmb{k} \in \mathbb{Z}^d} c_{\pmb{k}} (\tilde{\varphi}) \underbrace{ \left(\sum_{\pmb{\ell} \in I_{\pmb{n}}} g_{\pmb{\ell}} \mathrm{e}^{2 \pi \mathrm{i} \ \frac{1}{\pmb{n}} \odot (\pmb{k} \cdot \pmb{\ell})} \right)}_{ \eqqcolon \hat{g}_{\pmb{k}} \mathrm{e}^{-2 \pi \mathrm{i} \ \pmb{k} \cdot \pmb{x}}} \\ & = \sum_{\pmb{k} \in I_{\pmb{n}}} c_{\pmb{k}}(\tilde{\varphi}) \hat{g}_{\pmb{k}} \mathrm{e}^{-2 \pi \mathrm{i} \ \pmb{k} \cdot \pmb{x} } + \sum_{\pmb{r} \in \mathbb{Z}^d \setminus \{ \pmb{0} \}} \sum_{\pmb{k} \in I_{\pmb{n}}} c_{\pmb{k}}(\tilde{\varphi}) \hat{g}_{\pmb{k}} \mathrm{e}^{-2 \pi \mathrm{i} \ (\pmb{k} + \pmb{n} \odot \pmb{r})\cdot \pmb{x} }. \end{aligned}\]

Furthermore, a comparison of the equation above and $f$ suggests the choice

\[ \hat{g}_{\pmb{k}} = \begin{cases} \frac{\hat{f}_{\pmb{k}}}{c_{\pmb{k}}(\tilde{\varphi})} \quad &\text{for } \pmb{k} \in I_{\pmb{N}} \\ 0 &\text{for } \pmb{k} \in I_{\pmb{n}} \setminus I_{\pmb{N}} \end{cases}.\]

Now, we are able to compute $g_{\pmb{\ell}}$ by an FFT of size $n_1 \times n_2 \times \cdots \times n_d$. The error of the first approximation $f \approx s_1$ is called aliasing error. For further analysis, we refer to Chapter 7 in [Plonka, Potts, Steidl, Tasche, 2018].
We obtain an approximation $\psi$ for $\varphi$ if the window function is well localized in the spatial domain by the truncation

\[ \psi(\pmb{x}) \coloneqq \varphi(\pmb{x}) \mathbb{1}_{\times_{i = 1}^d [-m/n_i,m/n_i]}(\pmb{x})\]

with a chosen window size $m \ll \min_{i\in\{1,2,\dots,d\}}\{n_i\}, m \in \mathbb{N}$. Following the same scheme as above, we can use the periodization

\[ \tilde{\psi}(\pmb{x}) = \sum_{\pmb{r} \in \mathbb{Z}^d} \psi(\pmb{x}+\pmb{r}) \]

and look at the linear combination

\[ s(\pmb{x}_j) \coloneqq \sum_{\pmb{\ell} \in I_{\pmb{n}}} g_{\pmb{\ell}} \tilde{\psi} \left( \pmb{x}_j - \frac{1}{\pmb{n}} \odot \pmb{\ell} \right).\]

The following calculations show that $s \approx s_1$

\[\begin{aligned} s(\pmb{x}_j) &= \sum_{\pmb{\ell} \in I_{\pmb{n}}} g_{\pmb{\ell}} \sum_{\pmb{r} \in \mathbb{Z}^d} \psi \left( \pmb{x}_j - \frac{1}{\pmb{n}} \odot \pmb{\ell} + \pmb{r} \right) \\ &= \sum_{\pmb{\ell} \in I_{\pmb{n}}} g_{\pmb{\ell}} \sum_{\pmb{r} \in \mathbb{Z}^d} \varphi \left(\pmb{x}_j - \frac{1}{\pmb{n}} \odot \pmb{\ell} + \pmb{r} \right) \mathbb{1}_{\times_{i = 1}^d [-m/n_i,m/n_i]}(\pmb{x}_j) \\ &= \sum_{\ell \in I_n} g_{\pmb{\ell}} \mathbb{1}_{\times_{i = 1}^d [-m/n_i,m/n_i]}(\pmb{x}_j) \tilde{\varphi} \left(\pmb{x}_j - \frac{1}{\pmb{n}} \odot \pmb{\ell} \right) \\ &= \sum_{\pmb{\ell} \in I_{\pmb{n},m} (\pmb{x}_j)} g_{\pmb{\ell}} \tilde{\varphi} \left(\pmb{x}_j - \frac{1}{\pmb{n}} \odot \pmb{\ell} \right) \end{aligned}\]

with the index set

\[ I_{\pmb{n},m}(\pmb{x}_j) = \{ \pmb{\ell} \in I_{\pmb{n}} \colon \pmb{n} \odot \pmb{x}_j - m \pmb{1} \leq \pmb{\ell} \leq \pmb{n} \odot \pmb{x}_j +m \pmb{1} \}\]

for a fixed node $\pmb{x}_j$. This is motivated by

\[ -\frac{m}{n_i} \leq \left( \pmb{x}_j \right)_i \leq \frac{m}{n_i} \]

in order to ensure that $\pmb{x}_j$ is within the support. This second approximation error is called the truncation error. Summarizing, we have $f \approx s_1 \approx s$.

Pseudocode

Input: $N, M \in \mathbb{N}, \, \sigma > 1, \, m \in \mathbb{N}, \, x_j \in \mathbb{T}^d \text{ for } j \in I^{1}_{M}, \, f_k \in \mathbb{C} \text{ for } k \in I^{d}_{N}.$

Precomputation: Compute the nonzero Fourier coefficients $c_k(\tilde{\varphi}) \text{ for all } k \in I^{d}_{N}.$ Compute the values $\tilde{\psi} (x_j -\frac{\pi \pmb{l}}{\sigma N}) \text{ for } j \in I^{1}_{M} \text{ and } \pmb{l} \in I_{\sigma N, \ m} (x_j).$

Set $\hat{g}_k \coloneqq \hat{f}_k / c_k(\tilde{\varphi}) \text{ for } k \in I^{d}_{N}.$

Compute $g_{\pmb{l}} = \frac{1}{(\sigma N)^d} \sum_{k \in I^{d}_{N}} \hat{g}_k e^{2 \mathrm{i} \ k \cdot \pmb{l} / (\sigma N)}, \quad \pmb{l} \in I^{d}_{\sigma N}$ using a d-variate FFT.

Compute $s(x_j) \coloneqq \sum_{\pmb{l} \in I_{\sigma N, \ m} (x_j)} g_{\pmb{l}} \ \tilde{\psi}(x_j -\frac{\pi \pmb{l}}{\sigma N}), \quad \in I^{1}_{M}$.

Output: $s(x_j), \, j \in I^{1}_{M}$, approximate values of $f(x_j)$.

Computational cost: $\mathcal{O}(N^d \log{N} + m^d \ M)$

Adjoint Problem

Using the transposed index set

\[ I_{\pmb{n},m}^\top(\pmb{\ell}) = \{ j= 0, 1, \ldots, M-1 : \pmb{\ell} - m\pmb{1} \leq \pmb{n} \odot \pmb{x}_j \leq \pmb{\ell} + m \pmb{1} \},\]

we obtain the adjoint NFFT algorithm for the fast evaluation the adjoint problem

\[ \hat{h}_{\pmb{k}} = \sum_{j = 0}^{M-1} f_j \mathrm{e}^{2 \pi \ \mathrm{i} \ \pmb{k} \ \pmb{x}_j}, \pmb{k} \in I_{\pmb{N}}\]

for given coefficients $f_j \in \mathbb{C}$.

Functions

NFFT3.nfft_finalize_plan — Function

nfft_finalize_plan(P)

destroys a NFFT plan structure.

Input

P - a NFFT plan structure.

See also

NFFT{D}, nfft_init

NFFT3.nfft_init — Function

nfft_init(P)

intialises a transform plan.

Input

P - a NFFT plan structure.

See also

NFFT{D}, nfft_finalize_plan

NFFT3.nfft_trafo — Function

nfft_trafo(P)

computes a NFFT.

Input

P - a NFFT plan structure.

See also

NFFT{D}, nfft_trafo_direct

NFFT3.nfft_adjoint — Function

nfft_adjoint(P)

computes an adjoint NFFT.

Input

P - a NFFT plan structure.

See also

NFFT{D}, nfft_adjoint_direct

NFFT3.nfft_trafo_direct — Function

nfft_trafo_direct(P)

computes a NFFT.

Input

P - a NFFT plan structure.

See also

NFFT{D}, trafo

NFFT3.nfft_adjoint_direct — Function

nfft_adjoint_direct(P)

computes an adjoint NFFT.

Input

P - a NFFT plan structure.

See also

NFFT{D}, adjoint

Literature

[Keiner, Kunis, Potts, 2006] J. Keiner, S. Kunis, and D. Potts. Fast summation of radial functions on the sphere. Computing, 78:1–15, 2006.

[Plonka, Potts, Steidl, Tasche, 2018] G. Plonka, D. Potts, G. Steidl and M. Tasche. Numerical Fourier Analysis: Theory and Applications. Springer Nature Switzerland AG, 2018. doi: 10.1007/978-3-030-04306-3.

Schmischke2018

Schmischke, Michael Nonequispaced Fast Fourier Transform (NFFT) Interface for Julia. 2018 url: https://arxiv.org/abs/1810.09891

PlonkaPottsSteidelTasche2018

Plonka, Gerlind and Potts, Daniel and Steidl, Gabriele and Tasche, Manfred Numerical Fourier Analysis 2018

KeinerKunisPotts

J. Keiner, S. Kunis, and D. Potts NFFT 3.0, C subroutine library url: http://www.tu-chemnitz.de/~potts/nfft.